% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_inputs.R
\name{declineReasonInput}
\alias{declineReasonInput}
\title{Reason Patient Declined Participation}
\usage{
declineReasonInput(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
tagList
}
\description{
Reason Patient Declined Participation
}
