% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_inputs.R
\name{groupInputs}
\alias{groupInputs}
\title{Create a group of inputs}
\usage{
groupInputs(id, inputs)
}
\arguments{
\item{id}{namespace id}

\item{inputs}{tribble or data.frame with columns
\code{inputid} character string of inputId to use for this input,
\code{call} character string of function to call, and
\code{args}a list of additional arguments to pass to the function}
}
\value{
tagList
}
\description{
Create a group of inputs
}
