% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_inputs.R
\name{groupInputsServer}
\alias{groupInputsServer}
\title{Server Logic for a Group of Inputs}
\usage{
groupInputsServer(id, names)
}
\arguments{
\item{id}{namespace id}

\item{names}{names of input values in the group, should match
the names of \code{inputs} to \code{groupInputs} in order to capture
all input values, otherwise something will be missed}
}
\value{
list of values
}
\description{
Server Logic for a Group of Inputs
}
\details{
input an R6 object that has field names that match the
names of the inputs. When the trigger is triggered,
inputs will be saved by setting the matching R6 fields to the
input values.
}
