% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_hiddentabPanel.R
\name{hiddentabPanelUI}
\alias{hiddentabPanelUI}
\title{Module UI for a hidden tabsetPanel that is shown and hidden dynamically}
\usage{
hiddentabPanelUI(id, panels)
}
\arguments{
\item{id}{namespace id}

\item{panels}{named list of panels and their contents}
}
\value{
tagList of UI
}
\description{
Module UI for a hidden tabsetPanel that is shown and hidden dynamically
}
