% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_hiddentabPanel.R
\name{hideBtn}
\alias{hideBtn}
\title{Create button with ns id and label}
\usage{
hideBtn(id, label)
}
\arguments{
\item{id}{namespace id}

\item{label}{label to show in shiny}
}
\value{
actionButton
}
\description{
Create button with ns id and label
}
