% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_inputs.R
\name{ineligibleReasonInput}
\alias{ineligibleReasonInput}
\title{Input Reasons Patient Ineligible for Participation}
\usage{
ineligibleReasonInput(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
tagList
}
\description{
Input Reasons Patient Ineligible for Participation
}
