% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_inputs.R
\name{selectDiagnosis}
\alias{selectDiagnosis}
\title{Select Patients Diagnosis}
\usage{
selectDiagnosis(id)
}
\arguments{
\item{id}{namespaceid}
}
\value{
tagList
}
\description{
Select Patients Diagnosis
}
