% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_inputs.R
\name{selectOncologist}
\alias{selectOncologist}
\title{Select Patient's Oncologist}
\usage{
selectOncologist(id)
}
\arguments{
\item{id}{namespaceid}
}
\value{
none
}
\description{
Select Patient's Oncologist
}
