% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_select_patient.R
\name{selectPatient}
\alias{selectPatient}
\title{Select a Patient Module Input}
\usage{
selectPatient(inputId)
}
\arguments{
\item{inputId}{namespace id}
}
\value{
none
}
\description{
Select a Patient Module Input
}
