% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_select_patient.R
\name{selectPatientServer}
\alias{selectPatientServer}
\title{Select a Patient Module Server Logic}
\usage{
selectPatientServer(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
none
}
\description{
Select a Patient Module Server Logic
}
