% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_inputs.R
\name{selectPerson}
\alias{selectPerson}
\title{Select Person UI}
\usage{
selectPerson(id, label)
}
\arguments{
\item{id}{namespaceid}

\item{label}{label to display}
}
\value{
tagList
}
\description{
Select Person UI
}
