% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_hiddentabPanel.R
\name{showBtn}
\alias{showBtn}
\title{Create button with ns id and label}
\usage{
showBtn(id, label, icon = NULL)
}
\arguments{
\item{id}{namespace id}

\item{label}{label to show in shiny}

\item{icon}{optional parameter to pass an icon to the button}
}
\value{
actionButton
}
\description{
Create button with ns id and label
}
