% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_input}
\alias{wrap_input}
\title{Wrap a call to a shiny function with a namespace and call the argument}
\usage{
wrap_input(groupid, inputid, call, args)
}
\arguments{
\item{groupid}{id for the namespace}

\item{inputid}{id for the input}

\item{call}{function call}

\item{args}{list of other arguments}
}
\value{
shiny function
}
\description{
Wrap a call to a shiny function with a namespace and call the argument
}
