% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_hiddentabPanel.R
\name{wrapPanel}
\alias{wrapPanel}
\title{Wrap tab contents and its action button to hide panel}
\usage{
wrapPanel(id, contents, hidebutton)
}
\arguments{
\item{id}{panel name}

\item{contents}{panel contents}

\item{hidebutton}{actionButton type hideBtn}
}
\value{
tabPanel
}
\description{
Wrap tab contents and its action button to hide panel
}
