% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_inputs.R
\name{groupInputs}
\alias{groupInputs}
\title{Create a group of inputs}
\usage{
groupInputs(id, inputs)
}
\arguments{
\item{id}{namespace id}

\item{inputs}{tribble}
}
\value{
tagList
}
\description{
Create a group of inputs
}
