% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_hiddentabPanel.R
\name{hiddentabPanelServer}
\alias{hiddentabPanelServer}
\title{Hidden TabPanel Server Logic}
\usage{
hiddentabPanelServer(id, panelnames)
}
\arguments{
\item{id}{namespace id}

\item{panelnames}{character vector of names of panels}
}
\value{
none
}
\description{
Hidden TabPanel Server Logic
}
